# source/packages/contracts/control/openapi.yaml
openapi: 3.1.0
info:
  title: Personal Agent Control API
  version: 0.3.0
  description: |
    Control-plane contract for daemon clients.
    Unknown `/v1/*` routes return `404` problem+json envelopes with correlation headers.
servers:
  - url: http://127.0.0.1:7071
    description: Localhost TCP default binding
security:
  - bearerAuth: []
paths:
  /v1/tasks:
    post:
      operationId: submitTask
      summary: Submit a task.
      parameters:
        - $ref: '#/components/parameters/CorrelationID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskSubmitRequest'
      responses:
        '202':
          description: Accepted JSON response.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationID'

            X-PersonalAgent-API-Version:
              $ref: '#/components/headers/APIVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskSubmitResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/capabilities/smoke:
    get:
      operationId: capabilitySmoke
      summary: Transport smoke endpoint.
      parameters:
        - $ref: '#/components/parameters/CorrelationID'
      responses:
        '200':
          description: Successful JSON response.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationID'

            X-PersonalAgent-API-Version:
              $ref: '#/components/headers/APIVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilitySmokeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/realtime/ws:
    get:
      operationId: connectRealtimeWebSocket
      summary: Upgrade to WebSocket realtime stream.
      parameters:
        - $ref: '#/components/parameters/CorrelationID'
      responses:
        '101':
          description: Switching Protocols.
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  parameters:
    CorrelationID:
      name: X-Correlation-ID
      in: header
      required: false
      schema:
        type: string
      description: Optional correlation ID propagated across daemon responses/events.
...
  headers:
    CorrelationID:
      description: Correlation ID associated with this response.
      schema:
        type: string
    APIVersion:
      description: Daemon API contract version served by this endpoint.
      schema:
        type: string
        example: v1
...
  responses:
...
    BadRequest:
      description: Invalid request payload.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationID'
        X-PersonalAgent-API-Version:
          $ref: '#/components/headers/APIVersion'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid bearer token.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationID'
        X-PersonalAgent-API-Version:
          $ref: '#/components/headers/APIVersion'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    CapabilitySmokeResponse:
      type: object
      properties:
        daemon_version:
          type: string
        channels:
          type: array
          items:
            type: string
        connectors:
          type: array
          items:
            type: string
        healthy:
          type: boolean
        correlation_id:
          type: string
...
    TaskSubmitRequest:
      type: object
      required:
        - workspace_id
        - requested_by_actor_id
        - subject_principal_actor_id
        - title
      properties:
        workspace_id:
          type: string
        requested_by_actor_id:
          type: string
        subject_principal_actor_id:
          type: string
        title:
          type: string
        description:
          type: string
        task_class:
          type: string
    TaskSubmitResponse:
      type: object
      required:
        - task_id
        - run_id
        - state
        - correlation_id
      properties:
        task_id:
          type: string
        run_id:
          type: string
        state:
          type: string
        correlation_id:
          type: string
...
    ErrorResponse:
      type: object
      additionalProperties: true
      required:
        - error
        - correlation_id
        - type
        - title
        - status
        - detail
        - instance
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              description: |
                Optional domain-specific typed error metadata.
                For `service_not_configured` errors this includes `category`,
                `service` (`id`, `label`, `config_field`), `domain`, and `remediation`
                (`action`, `label`, `hint`).
        correlation_id:
          type: string
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string